+++ farm_rothamsted.post_update.php
@@ -5,6 +5,7 @@
  * Update hooks for farm_rothamsted.module.
  */
 
+use Drupal\views\Entity\View;
 use Symfony\Component\Yaml\Yaml;
 
 /**
@@ -79,4 +80,18 @@ function farm_rothamsted_post_update_create_rothamsted_asset_parent_action(&$san
   $config_path = \Drupal::service('extension.list.module')->getPath('farm_rothamsted') . "/config/optional/system.action.$action_id.yml";
   $data = Yaml::parseFile($config_path);
   \Drupal::configFactory()->getEditable("system.action.$action_id")->setData($data)->save(TRUE);
+}
+
+/**
+ * Create rothamsted_uncategorized_logs view.
+ */
+function farm_rothamsted_post_update_create_rothamsted_uncategorized_logs_view(&$sandbox = NULL) {
+
+  // Only create the view if views is enabled.
+  $view_id = 'rothamsted_uncategorized_logs';
+  if (\Drupal::moduleHandler()->moduleExists('views') && !View::load($view_id)) {
+    $config_path = \Drupal::service('extension.list.module')->getPath('farm_rothamsted') . "/config/optional/views.view.$view_id.yml";
+    $data = Yaml::parseFile($config_path);
+    \Drupal::configFactory()->getEditable("views.view.$view_id")->setData($data)->save(TRUE);
+  }
 }