+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Migrate contacts to a user reference field. [#322](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/322)
 - Replace field locations with location asset reference. [#314](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/314)
 - Save crop and variety in drilling and commercial quick forms. [#310](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/310)
+- Change plot assets to not default as locations. [#336](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/336)
 
 ## [2.8.0](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/milestone/13) 2022-10-28