+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add baseline and seed_multiplication plot types. [#333](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/333)
 - Add seed_dressing field to drilling logs. [#311](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/311)
 - Add separate views of experiment logs in secondary tabs. [#303](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/303)
+- Add calibration plot type. [#205](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/205)
 
 ### Changed